+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Bump MSRV from `1.71` to `1.85`.
 - Fixed EGL's robustness detection without extension, but on EGL 1.5.
 - Fixed building docs on docs.rs.
+- Fixed WGL's unreliable TRANSPARENT_ARB false negatives by returning None
 
 # Version 0.32.3